Move logic inside the SyncBlog#contentful method to separate classes
Now that we users have the chance to manipulate the mappings and create
more elaborated ones it's safer to use a blank object as a context
object.

For that I have create the TemplateContext object wich is basically a
BasicObject with some metaprogramming on it plus a helper to gets its
bindings.

Besides that I created a 'renderer' class to have better boundaries
between responsabilities.

Right now for simplicity both new classes are on the same file as
SyncBlog.